backend/requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==47.0.0
Django==6.0.4
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.1
gunicorn==26.0.0
mysqlclient==2.2.8
packaging==26.2
pillow==12.2.0
pycparser==3.0
PyMySQL==1.1.3
sqlparse==0.5.5
tzdata==2026.2
whitenoise==6.12.0
